+++ CHANGELOG.md
@@ -13,6 +13,12 @@
   * added `bovigo\assert\phpunit\PHPUnit_Framework_TestCase::assertStringEndsWith()`
   * added `bovigo\assert\predicate\doesNotEndWith()`
   * added `bovigo\assert\phpunit\PHPUnit_Framework_TestCase::assertStringEndsNotWith()`
+  * added `bovigo\assert\predicate\isNan()`
+  * added `bovigo\assert\phpunit\PHPUnit_Framework_TestCase::assertNan()`
+  * added `bovigo\assert\predicate\isFinite()`
+  * added `bovigo\assert\phpunit\PHPUnit_Framework_TestCase::assertFinite()`
+  * added `bovigo\assert\predicate\isInfinite()`
+  * added `bovigo\assert\phpunit\PHPUnit_Framework_TestCase::assertInfinite()`
 
 
 1.0.0 (2015-12-27)